disable WS pings and improve logging
The "disconnect loop" that I have observed seems to be related to the Grizzl-e sending invalid WebSocket "pongs" sometimes.

The disconnect is preceded by Grizzl-e sending a truncated pong, 1, 2 or 3 bytes when it should be 4 bytes:

PING 98 07 f9 00 [binary, 4 bytes]
< PONG 98 07 f9 [binary, 3 bytes]

PING b6 3c 19 00 [binary, 4 bytes]
< PONG b6 3c 19 [binary, 3 bytes]

When the invalid pong is received, the server disconnects. Seems likely to be a bug in TinyWebsockets.

Disabling WS pings means we'll need to rely on Heartbeat messages and long connection timeouts